Fix syntax

+++ bin/create_aws_cloudformation.sh
@@ -176,7 +176,7 @@ fi
 # Check if we have an existing AWS SSH key that has the correct name
 "$AWS" --profile "$AWS_PROFILE" ec2 describe-key-pairs --key-names "$BOSH_SSH_KEY_NAME" >/dev/null 2>&1 && AWS_KEY_EXISTS='true'
 
-if [ x"$REGENERATE_SSH_KEY" = x"true" -f "$BOSH_SSH_KEY_FILENAME" ]; then
+if [ x"$REGENERATE_SSH_KEY" = x"true" -a -f "$BOSH_SSH_KEY_FILENAME" ]; then
 	INFO 'Deleting local SSH key'
 	rm -f "$BOSH_SSH_KEY_FILENAME" "$BOSH_SSH_KEY_FILENAME.pub"
 fi
